Rework behaviour of Flip and Flashcard
Flip command should toggle the English phrase of a flashcard to hide or
show. The current behaviour of flip does not do this. Furthermore the
setIsFlipped method of Flashcard no longer makes Flashcard immutable,
which was the original intended design. This could cause future code
regression.

Changing the behaviour of Flip to toggle hiding for English phrases will
allow the in-app flashcard to behave more like a real-life flashcard.
Replacing the setIsFlipped method for Flashcard will also allow
Flashcard to regain its immutability.

Let's,
* make changes to the flip command to enable it to hide and show
English phrases upon invocation.
* replace setIsFlipped method with a getFlippedFlashcard method that
returns a copy of the current Flashcard but flipped.