Add state to slides in slideshow
When the user is in slideshow mode, they can now view the current
flashcard number, as well as the number of flashcards they have answered
in the current slideshow.

In addition, when a user answers a question, the slide will now continue
to show the answer after the user moves to other slides.